Core/Creature: Implemented negative levels for creatures having -1 as expansion.
Authored by Kinzcool. https://github.com/TrinityCore/TrinityCore/commit/e3b23c2513a651242d3c04e222ac05aa0330400c

+++ src/server/game/Entities/Creature/Creature.h
@@ -86,9 +86,9 @@ struct CreatureTemplate
     std::string  SubName;
     std::string  IconName;
     uint32  GossipMenuId;
-    uint8   minlevel;
-    uint8   maxlevel;
-    uint32  expansion;
+    int16   minlevel;
+    int16   maxlevel;
+    int32  expansion;
     uint32  expansionUnknown;                               // either 0 or 3, sent to the client / wdb
     uint32  faction_A;
     uint32  faction_H;
